refactor(filid): split fractal-tree into tree-builder + scanner organs (T1-B)
Split core/tree/fractal-tree/fractal-tree.ts (490 lines, 13 functions across
2 clear axes) into an organ-first structure. 23 consumers across src/core/
and src/index.ts — all preserved by the slim facade.

tree-builder/ (pure tree construction + navigation):
- build-fractal-tree.ts       — buildFractalTree + private findParentPath + NodeEntry type
- find-node.ts, get-ancestors.ts, get-descendants.ts
- get-fractals-under-organs.ts

scanner/ (filesystem scanning + framework detection):
- scan-project.ts             — main entry orchestrating sibling scanner files
- should-exclude.ts
- detect-frameworks.ts        — detectFrameworks + detectFrameworkReserved
- discover-directories.ts, collect-node-metadata.ts, correct-node-types.ts

fractal-tree.ts is now a 9-line slim facade with export * + explicit
export type { NodeEntry }.

- src/index.ts / src/core/index.ts: empty diff (23 consumers untouched)
- fractal-tree.ts: 490 -> 9 lines
- INTENT.md: 22 lines (<=50)
- bundle delta: max +4.03% (context-injector.mjs, well under 5% gate)
- tests: 846/850 (4 pre-existing context-injector failures)
