+++ packages/filid/src/core/tree/fractal-tree/INTENT.md
@@ -4,6 +4,12 @@
 
 프랙탈 트리 구축, 스캔, 노드 탐색.
 
+## Structure
+
+- `tree-builder/` — 트리 구축 및 탐색 (buildFractalTree, findNode, getAncestors, getDescendants, getFractalsUnderOrgans)
+- `scanner/` — 파일시스템 스캔 및 프레임워크 감지 (scanProject, shouldExclude, detectFrameworks, discoverDirectories, collectNodeMetadata, correctNodeTypes)
+- `fractal-tree.ts` — slim facade (tree-builder/ + scanner/ re-export)
+
 ## Boundaries
 
 ### Always do
